Added IsProperty() check to result in DefineOrRedefineAccessor - if no check this causes the debug check to fails since type() asserts that the lookup_type != NOT_FOUND. This does not change any functionality since we explicitly checked if it was one of the three types that we need to delete. Also changed defineProperties in v8natives to actually return the object being defined (as specified by spec)
Review URL: http://codereview.chromium.org/572003

git-svn-id: http://v8.googlecode.com/svn/branches/bleeding_edge@3794 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ src/runtime.cc
@@ -2906,10 +2906,11 @@ static Object* Runtime_DefineOrRedefineAccessorProperty(Arguments args) {
   // If an existing property is either FIELD, NORMAL or CONSTANT_FUNCTION
   // delete it to avoid running into trouble in DefineAccessor, which
   // handles this incorrectly if the property is readonly (does nothing)
-  if (result.type() == FIELD || result.type() == NORMAL
-      || result.type() == CONSTANT_FUNCTION)
+  if (result.IsValid() &&
+      (result.type() == FIELD || result.type() == NORMAL
+       || result.type() == CONSTANT_FUNCTION)) {
     obj->DeleteProperty(name, JSObject::NORMAL_DELETION);
-
+  }
   return obj->DefineAccessor(name, flag_setter->value() == 0, fun, attr);
 }
